+++ doc/changelog.rst
@@ -52,6 +52,12 @@ PyMongo 4.9 brings a number of improvements including:
         File "/Users/<user>/my_file.py", line 8, in <module>``
             client = MongoClient()
     Call MongoClient.close() to safely shut down your client and free up resources.
+- The default value for ``connect`` in ``MongoClient`` is changed to ``False`` when running on
+  unction-as-a-service (FaaS) like AWS Lambda, Google Cloud Functions, and Microsoft Azure Functions.
+  On some FaaS systems, there is a ``fork()`` operation at function
+  startup.  By delaying the connection to the first operation, we avoid a deadlock.  See
+  `Is PyMongo Fork-Safe`_ for more information.
+
 
 Issues Resolved
 ...............
@@ -59,6 +65,7 @@ Issues Resolved
 See the `PyMongo 4.9 release notes in JIRA`_ for the list of resolved issues
 in this release.
 
+.. _Is PyMongo Fork-Safe : https://www.mongodb.com/docs/languages/python/pymongo-driver/current/faq/#is-pymongo-fork-safe-
 .. _PyMongo 4.9 release notes in JIRA: https://jira.mongodb.org/secure/ReleaseNote.jspa?projectId=10004&version=39940